requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via starlette
beautifulsoup4==4.12.3
    # via bs4
bs4==0.0.2
    # via -r requirements.in
cachetools==5.5.0
    # via google-auth
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
decorator==5.1.1
    # via retry
fastapi==0.112.1
    # via -r requirements.in
google-api-core==2.19.1
    # via google-api-python-client
google-api-python-client==2.141.0
    # via pygsheets
google-auth==2.34.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.1
    # via
    #   -r requirements.in
    #   pygsheets
googleapis-common-protos==1.63.2
    # via google-api-core
h11==0.14.0
    # via uvicorn
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.7
    # via
    #   anyio
    #   requests
lxml==5.3.0
    # via -r requirements.in
numpy==2.1.0
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
pandas==2.2.2
    # via -r requirements.in
pillow-avif-plugin==1.4.6
    # via -r requirements.in
proto-plus==1.24.0
    # via google-api-core
protobuf==5.27.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
py==1.11.0
    # via retry
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pydantic==2.8.2
    # via fastapi
pydantic-core==2.20.1
    # via pydantic
pygsheets==2.0.6
    # via -r requirements.in
pyparsing==3.1.2
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.1
    # via pandas
requests==2.32.3
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
retry==0.9.2
    # via -r requirements.in
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
starlette==0.38.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.2
    # via requests
uvicorn==0.30.6
    # via -r requirements.in
